make tfgen build_sdks

+++ sdk/dotnet/Inputs/RulesetRuleActionParametersFromValueGetArgs.cs
@@ -13,19 +13,19 @@ namespace Pulumi.Cloudflare.Inputs
     public sealed class RulesetRuleActionParametersFromValueGetArgs : global::Pulumi.ResourceArgs
     {
         /// <summary>
-        /// Preserve query string for redirect URL.
+        /// Keep the query string of the original request.
         /// </summary>
         [Input("preserveQueryString")]
         public Input<bool>? PreserveQueryString { get; set; }
 
         /// <summary>
-        /// Status code for redirect.
+        /// The status code to be used for the redirect.
         /// </summary>
         [Input("statusCode")]
-        public Input<int>? StatusCode { get; set; }
+        public Input<double>? StatusCode { get; set; }
 
         /// <summary>
-        /// Target URL for redirect.
+        /// The URL to redirect the request to.
         /// </summary>
         [Input("targetUrl")]
         public Input<Inputs.RulesetRuleActionParametersFromValueTargetUrlGetArgs>? TargetUrl { get; set; }
